Update tests to include ubuntu
