+++ molecule/resources/tests/test_default.py
@@ -1,10 +1,13 @@
 import os
 import pytest
 import testinfra.utils.ansible_runner
+from time import time
 
 testinfra_hosts = testinfra.utils.ansible_runner.AnsibleRunner(
     os.environ['MOLECULE_INVENTORY_FILE']).get_hosts('all')
 
+# Ubuntu sudo doesn't set HOME so it tries to write to /root
+ENV = 'OMERO_USERDIR=/home/data-importer/omero-{}'.format(time())
 OMERO = '/opt/omero/server/OMERO.server/bin/omero'
 OMERO_LOGIN = '-C -s localhost -u root -w omero'
 
@@ -16,7 +19,7 @@ def test_service_running_and_enabled(host):
 
 def test_omero_root_login(host):
     with host.sudo('data-importer'):
-        host.check_output('%s login %s' % (OMERO, OMERO_LOGIN))
+        host.check_output('%s %s login %s' % (ENV, OMERO, OMERO_LOGIN))
 
 
 @pytest.mark.parametrize("key,value", [
@@ -27,7 +30,7 @@ def test_omero_root_login(host):
 ])
 def test_omero_server_config(host, key, value):
     with host.sudo('omero-server'):
-        cfg = host.check_output("%s config get %s", OMERO, key)
+        cfg = host.check_output('%s %s config get %s' % (ENV, OMERO, key))
     assert cfg == value
 
 
@@ -45,3 +48,29 @@ def test_omero_managedrepo(host):
     assert d.user == 'omero-server'
     assert d.group == 'importer'
     assert d.mode == 0o2775
+
+
+def test_inplace_import(host):
+    fake_file = '/data/import/test.fake'
+    with host.sudo('data-importer'):
+        outimport = host.check_output(
+            '%s %s %s import --skip=upgrade --transfer=ln_s %s' %
+            (ENV, OMERO, OMERO_LOGIN, fake_file))
+
+    imageid = int(outimport.split(':', 1)[1])
+    assert imageid
+
+    query = ('SELECT concat(ofile.path, ofile.name) '
+             'FROM FilesetEntry AS fse '
+             'JOIN fse.fileset AS fileset '
+             'JOIN fse.originalFile AS ofile '
+             'JOIN fileset.images AS image '
+             'WHERE image.id = %d' % imageid)
+    with host.sudo('data-importer'):
+        outhql = host.check_output(
+            '%s %s %s hql -q --style plain "%s"' %
+            (ENV, OMERO, OMERO_LOGIN, query))
+
+    f = host.file('/OMERO/ManagedRepository/%s' % outhql.split(',', 1)[1])
+    assert f.is_symlink
+    assert f.linked_to == fake_file
